Make wheel class reusable
This can be used by other parts of the project.

Fixes: https://github.com/stanislavlevin/pyproject_installer/issues/21
